pyproject.toml
[project]
name = "mcp-gamecube-bridge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.2.0",
    "pillow>=11.2.1",
    "pygame>=2.6.1",
    "pyserial>=3.5",
]

[dependency-groups]
dev = [
    "ruff>=0.11.6",
]
